pyproject.toml
[project]
name = "LESSPayne"
version = "0.1"
description = "LESSPayne"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
    "numpy==1.21.0",
    "matplotlib==3.1.3",
    "pyside2==5.14.2",
    "astropy",
    "ipython",
    "requests",
    "jupyter",
    "psutil",
    "numpy",
    "pyyaml",
    "scipy>=0.14.0",
    "six",
]
authors = [
    {name = "Alex Ji", email = "alexji@uchicago.edu"},
    {name = "Andy Casey", email = "andy.casey@monash.edu"},
    {name = "Jennifer Sobeck"},
    {name = "Riley Thai", email = "rileythai@proton.me"},
]
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/alexji/LESSPayne"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
mypkg = ["*.moog", "*.txt","*.fits","*.yaml", "*.pkl"]

[project.scripts]
lesspayne = "lesspayne.cli:run"

[project.gui-scripts]
smhr = "lesspayne.smh.gui.__main__:run_gui"
